requirements.txt
streamlit
pandas
schedule
pytz
instagrapi
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
requests
selenium
